updated ms repo

+++ milestones/summer24--sft221-nbb-2/SourceCode/Milestone/Milestone_main/data.h
@@ -2,6 +2,9 @@
 #define DATA_H
 #include "mapping.h"
 
+
+
+
 /**
  * Struct for a Package or Shipment
  * @param Point destination - points on the map corresponding to package destination
@@ -25,7 +28,7 @@ struct Package {
 struct Truck {
 	struct Point* currentPosition;
     double currentWeight;
-    double currentVolume;
+    int currentVolume;
     char routeSymbol;
     struct Route* route;
     int routeLength;
